requirements.txt
Flask
Flask-RESTful
Flask-Migrate
SQLAlchemy
psycopg2-binary
pydantic
pandas
requests
python-dotenv
email-validator
pytest
gunicorn
alembic
flask db upgrade


